test: split GridViewIT into multiple test files
