+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/pom.xml
@@ -158,6 +158,11 @@
             <version>${flow.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.vaadin</groupId>
+            <artifactId>flow-client</artifactId>
+            <version>${flow.version}</version>
+        </dependency>
+        <dependency>
             <groupId>javax.servlet</groupId>
             <artifactId>javax.servlet-api</artifactId>
             <scope>provided</scope>
