+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/it/GridViewHeightByRowsIT.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2000-2022 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.grid.it;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+import com.vaadin.flow.component.grid.testbench.GridElement;
+import com.vaadin.flow.testutil.TestPath;
+import com.vaadin.tests.AbstractComponentIT;
+
+@TestPath("vaadin-grid-it-demo/height-by-rows")
+public class GridViewHeightByRowsIT extends AbstractComponentIT {
+
+    @Before
+    public void init() {
+        open();
+    }
+
+
+    @Test
+    public void heightByRows_allRowsAreFetched() {
+        GridElement grid = $(GridElement.class).id("grid-height-by-rows");
+        scrollToElement(grid);
+        waitUntil(driver -> grid.getRowCount() == 50);
+
+        Assert.assertEquals("Grid should have heightByRows set to true", "true",
+                grid.getAttribute("allRowsVisible"));
+    }
+}
